requirements.txt
flask
python-dotenv
fyers-apiv3
pytz